% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unpack.df.columns.R
\name{unpack.df.columns}
\alias{unpack.df.columns}
\title{unpack.df.columns}
\usage{
unpack.df.columns(df)
}
\arguments{
\item{df}{the data.frame to unpack}
}
\value{
the unpacked, flattened data.frame
}
\description{
Function for unpacking dataframes where columns also contain dataframes. It returns a dataframe with appropriately-named atomic columns.
}
\examples{
unpack.df.columns(df)
}
